+++ api/src/main/java/com/tinqinacademy/commentsservice/api/base/OperationProcessor.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.commentsservice.api.base;
 
-import com.tinqinacademy.commentsservice.api.error.Errors;
+import com.tinqinacademy.commentsservice.api.exceptions.Errors;
 import io.vavr.control.Either;
 
 public interface OperationProcessor<I extends  OperationInput, O extends OperationOutput > {
