+++ api/src/main/java/com/tinqinacademy/commentsservice/api/exceptions/Errors.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.commentsservice.api.error;
+package com.tinqinacademy.commentsservice.api.exceptions;
 
 import lombok.*;
 import org.springframework.http.HttpStatus;
