+++ api/src/main/java/com/tinqinacademy/commentsservice/api/exceptions/custom/CustomException.java
@@ -1,4 +1,4 @@
-package com.tinqinacademy.commentsservice.core.exceptions.custom;
+package com.tinqinacademy.commentsservice.api.exceptions.custom;
 
 import org.springframework.http.HttpStatus;
