+++ api/src/main/java/com/tinqinacademy/commentsservice/api/exceptions/custom/ViolationsException.java
@@ -1,6 +1,6 @@
-package com.tinqinacademy.commentsservice.core.exceptions.custom;
+package com.tinqinacademy.commentsservice.api.exceptions.custom;
 
-import com.tinqinacademy.commentsservice.api.error.Error;
+import com.tinqinacademy.commentsservice.api.exceptions.Error;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
