+++ core/src/main/java/com/tinqinacademy/commentsservice/core/exceptions/ExceptionService.java
@@ -1,6 +1,6 @@
 package com.tinqinacademy.commentsservice.core.exceptions;
 
-import com.tinqinacademy.commentsservice.api.error.Errors;
+import com.tinqinacademy.commentsservice.api.exceptions.Errors;
 public interface ExceptionService {
     Errors handle(Throwable throwable);
 }
