+++ core/src/main/java/com/tinqinacademy/commentsservice/core/operations/system/AdminEditCommentForRoomOperationProcessor.java
@@ -1,11 +1,11 @@
 package com.tinqinacademy.commentsservice.core.operations.system;
 
-import com.tinqinacademy.commentsservice.api.error.Errors;
+import com.tinqinacademy.commentsservice.api.exceptions.Errors;
 import com.tinqinacademy.commentsservice.api.operations.system.editcommentforroom.AdminEditCommentForRoomInput;
 import com.tinqinacademy.commentsservice.api.operations.system.editcommentforroom.AdminEditCommentForRoomOperation;
 import com.tinqinacademy.commentsservice.api.operations.system.editcommentforroom.AdminEditCommentForRoomOutput;
 import com.tinqinacademy.commentsservice.core.exceptions.ExceptionService;
-import com.tinqinacademy.commentsservice.core.exceptions.custom.NotFoundException;
+import com.tinqinacademy.commentsservice.api.exceptions.custom.NotFoundException;
 import com.tinqinacademy.commentsservice.core.operations.BaseOperationProcessor;
 import com.tinqinacademy.commentsservice.core.utils.LoggingUtils;
 import com.tinqinacademy.commentsservice.persistence.model.entity.Comment;
